commit #21
Action:
halaman menu, detail menu itu diubah jadi detail pesanan (done)
filter tagihan belum di eksekusi (done)
validasi tanggal booking, tidak boleh kurang dari hari ini (done)
alert hapus booking harus diperbaiki ketika konfirmasi (done)
tampilan kolom nama kasir pada manajemen bahan (done)
tampilan menu favorit (done)
tampilan penjualan harian (done)

+++ database/migrations/2023_03_01_110657_add_relation_transactions_transaction_details.php
@@ -23,8 +23,11 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('transactions', function (Blueprint $table) {
-            $table->dropForeign(['transaction_details_id']);
-            $table->dropColumn('transaction_details_id');
+            // check if column exists
+            if (Schema::hasColumn('transactions', 'transaction_details_id')) {
+                $table->dropForeign('transactions_transaction_details_id_foreign');
+                $table->dropColumn('transaction_details_id');
+            }
         });
     }
 };
